requirements.txt
aiocron==1.8
aiohttp==3.11.8
Flask==3.1.0
openai==1.55.3
pydantic==2.10.2
python-dotenv==1.0.1
PyYAML==6.0.2
Requests==2.32.3
supabase==2.10.0
